refactor(locks): merge locksAuth + locksContent into a single locks domain per layer (#2026)
