--- src/core/controllers/locksContent/locksContent.test.ts
@@ -1,130 +1,0 @@
-import type { Session as LocksSdkSession } from '@pubky/locks-sdk';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { isAppError } from '@/libs/error/error.utils';
-import type { TGuardedResource } from '@/services/locksContent/locksContent.types';
-import { useLocksAuthStore } from '@/stores/locksAuth/locksAuth.store';
-import { locksAuthInitialState } from '@/stores/locksAuth/locksAuth.types';
-import { asOpaque } from '@/test-utils/type-assertions';
-import { LocksContentController } from './locksContent';
-
-const mocks = vi.hoisted(() => ({
-  registerGuardedResource: vi.fn(),
-  createContentLock: vi.fn(),
-  randomUUID: vi.fn(),
-}));
-
-vi.mock('@/application/locksContent/locksContent', () => ({
-  LocksContentApplication: {
-    registerGuardedResource: mocks.registerGuardedResource,
-    createContentLock: mocks.createContentLock,
-  },
-}));
-
-const fakeSession = asOpaque<LocksSdkSession>({ id: 'locks-session', lockServer: () => 'lockpubky' });
-
-const file = (contentType = 'application/json') => ({ contentType, bytes: new Uint8Array([1]) });
-const descriptor = (path: string) => ({ path, hash: 'HASH', content_type: 'application/json', size: 1 });
-/** Default builder: ignores the attachment paths and returns a fixed post JSON. */
-const buildPost = () => file();
-
-describe('LocksContentController', () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    let minted = 0;
-    mocks.randomUUID.mockImplementation(() => `id-${++minted}`);
-    vi.stubGlobal('crypto', { ...globalThis.crypto, randomUUID: mocks.randomUUID });
-    mocks.registerGuardedResource.mockImplementation(({ path }: { path: string }) => ({
-      resource: descriptor(path),
-      creator: 'pubkybob',
-    }));
-    mocks.createContentLock.mockResolvedValue({
-      lock_id: 'LOCK1',
-      content_lock_path: '/pub/locks.app/LOCK1.json',
-      creator: 'pubkybob',
-    });
-    useLocksAuthStore.setState(locksAuthInitialState);
-    useLocksAuthStore.getState().init({ session: fakeSession, secret: 'secret-abc' });
-  });
-
-  it('uploads the attachments first, then the post built from their paths', async () => {
-    const seen: TGuardedResource[][] = [];
-    let seenOwner: string | undefined;
-    const result = await LocksContentController.createLockContent({
-      attachments: [file('image/png'), file('video/mp4')],
-      buildPost: (attachmentResources, ownerPubky) => {
-        seen.push(attachmentResources);
-        seenOwner = ownerPubky;
-        return file();
-      },
-    });
-
-    // Attachments upload first (id-1, id-2); the post is uploaded last (id-3).
-    expect(mocks.registerGuardedResource.mock.calls.map(([params]) => params.path)).toEqual(['id-1', 'id-2', 'id-3']);
-    // The builder sees the attachment descriptors, so the post can reference their paths.
-    expect(seen).toEqual([[descriptor('id-1'), descriptor('id-2')]]);
-    // ...and the owner the bytes landed on, so those references point at the right account.
-    expect(seenOwner).toBe('pubkybob');
-
-    expect(mocks.createContentLock).toHaveBeenCalledWith(
-      expect.objectContaining({
-        session: fakeSession,
-        primaryResource: descriptor('id-3'),
-        secondaryResources: [descriptor('id-1'), descriptor('id-2')],
-        lockServer: { override: 'lockpubky' },
-      }),
-    );
-    expect(result).toEqual({ lock_id: 'LOCK1', content_lock_path: '/pub/locks.app/LOCK1.json', creator: 'pubkybob' });
-  });
-
-  it('gives identical files distinct paths, so neither overwrites the other', async () => {
-    await LocksContentController.createLockContent({
-      attachments: [file('image/png'), file('image/png')],
-      buildPost,
-    });
-
-    const paths = mocks.registerGuardedResource.mock.calls.map(([params]) => params.path);
-    expect(new Set(paths).size).toBe(paths.length);
-  });
-
-  it('builds the post with an empty list when there are no attachments', async () => {
-    const builder = vi.fn(() => file());
-
-    await LocksContentController.createLockContent({ buildPost: builder });
-
-    expect(builder).toHaveBeenCalledWith([], undefined);
-    expect(mocks.registerGuardedResource).toHaveBeenCalledTimes(1);
-    expect(mocks.createContentLock).toHaveBeenCalledWith(expect.objectContaining({ secondaryResources: [] }));
-  });
-
-  it('sends the placeholder dev-static criterion and lock logic', async () => {
-    await LocksContentController.createLockContent({ buildPost });
-
-    const [params] = mocks.createContentLock.mock.calls[0];
-    expect(params.criteria).toEqual([
-      { criterion_id: 'criterion-1', verifier_type: 'dev-static', params: { satisfied: true } },
-    ]);
-    expect(params.lockLogic).toEqual({ type: 'all', criteria: ['criterion-1'] });
-    expect(params.accessPolicy).toEqual({ requested_credential_ttl_seconds: 900 });
-  });
-
-  it('rejects when there is no Locks session', async () => {
-    useLocksAuthStore.getState().reset();
-
-    const error = await LocksContentController.createLockContent({ buildPost }).catch((caught: unknown) => caught);
-
-    expect(isAppError(error)).toBe(true);
-    expect(mocks.registerGuardedResource).not.toHaveBeenCalled();
-  });
-
-  it('does not build the post or create the lock when an attachment upload fails', async () => {
-    mocks.registerGuardedResource.mockRejectedValueOnce(new Error('upload failed'));
-    const builder = vi.fn(() => file());
-
-    await expect(
-      LocksContentController.createLockContent({ attachments: [file('image/png')], buildPost: builder }),
-    ).rejects.toThrow('upload failed');
-
-    expect(builder).not.toHaveBeenCalled();
-    expect(mocks.createContentLock).not.toHaveBeenCalled();
-  });
-});
