+++ src/core/services/locks/locks.ts
@@ -1,4 +1,14 @@
-import { CreateContentLockRequestBuilder, RegisterGuardedResourceOptions } from '@pubky/locks-sdk';
+import {
+  ConnectUrlOptions,
+  CreateContentLockRequestBuilder,
+  ExchangeFrontendSessionCodeOptions,
+  Locks,
+  LocksOptions,
+  RegisterGuardedResourceOptions,
+  type Session as LocksSdkSession,
+  SetLockServicePointerOptions,
+} from '@pubky/locks-sdk';
+import { getHomeserver, getPkarrRelays, getTestnet } from '@/config/network';
 import { AuthErrorCode } from '@/libs/error/error.codes';
 import { Err } from '@/libs/error/error.factories';
 import { ErrorService } from '@/libs/error/error.types';
@@ -7,10 +17,18 @@ import { ensureLocksSdkReady } from '@/libs/locks/locksSdk';
 import type {
   TCreateContentLockParams,
   TCreateContentLockResult,
+  TExchangeSessionCodeParams,
+  TGenerateConnectUrlParams,
   TGuardedResource,
+  TLocksSessionResult,
   TRegisterGuardedResourceParams,
   TRegisterGuardedResourceResult,
-} from './locksContent.types';
+  TRestoreLocksSessionParams,
+} from './locks.types';
+
+/** Opt-in flag telling the Lock Server `/connect` shell to deliver the code via postMessage
+ * instead of redirecting back to `returnTo`. */
+const DELIVERY_POSTMESSAGE = 'postmessage';
 
 /**
  * The SDK reports HTTP failures as `Lock Server request failed with HTTP <status>` and exposes no
@@ -30,15 +48,95 @@ function toLocksError(error: unknown, operation: string) {
 }
 
 /**
- * IO boundary for publishing locked content via the locks-sdk creator API. Every call goes through
- * a live `session` (bound to its Lock Server), so — unlike `LocksAuthService` — this service needs
- * no network config of its own.
+ * IO boundary for the Lock Server via the locks-sdk.
  *
- * Two calls compose one lock: N × `registerGuardedResource` (raw bytes per file) then
+ * Auth methods open sessions, so they build a Locks client from the app's network config
+ * (pkarr relays / testnet homeserver, like `HomeserverService`). Content methods publish locked
+ * content through a live `session` (bound to its Lock Server), so they need no network config —
+ * two calls compose one lock: N × `registerGuardedResource` (raw bytes per file) then
  * 1 × `createContentLock` (bundles the returned descriptors + unlock rules).
  */
-export class LocksContentService {
+export class LocksService {
   private constructor() {} // Prevent instantiation
+
+  /** Builds a Locks client bound to the given Lock Server, using the app's network config. */
+  private static buildLocksClient(lockServerPubky: string): Locks {
+    const options = new LocksOptions();
+    for (const relay of getPkarrRelays()) {
+      options.addPkarrRelay(relay);
+    }
+    if (getTestnet()) {
+      options.setLocalTestnetHomeserver(getHomeserver());
+    }
+    return Locks.forServerWithOptions(lockServerPubky, options);
+  }
+
+  /**
+   * Builds the Lock-Server-hosted `/connect` URL (resolves the server via pkarr) and opts into
+   * postMessage delivery so the shell posts the code to the parent instead of redirecting.
+   */
+  static async generateConnectUrl({ lockServerPubky, returnTo, state }: TGenerateConnectUrlParams): Promise<string> {
+    try {
+      await ensureLocksSdkReady();
+      const locks = this.buildLocksClient(lockServerPubky);
+      const url = await locks.createConnectUrl(new ConnectUrlOptions(returnTo, state));
+      const withDelivery = new URL(url);
+      withDelivery.searchParams.set('delivery', DELIVERY_POSTMESSAGE);
+      return withDelivery.toString();
+    } catch (error) {
+      throw toAppError(error, ErrorService.Locks, 'LocksService.generateConnectUrl');
+    }
+  }
+
+  /** Exchanges the one-time callback code for a Locks session. */
+  static async exchangeSessionCode({
+    lockServerPubky,
+    code,
+    state,
+  }: TExchangeSessionCodeParams): Promise<TLocksSessionResult> {
+    try {
+      await ensureLocksSdkReady();
+      const locks = this.buildLocksClient(lockServerPubky);
+      const session = await locks.exchangeFrontendSessionCode(new ExchangeFrontendSessionCodeOptions(code, state));
+      // secret is freshly minted here — export it so the caller can persist it.
+      return { session, secret: session.exportSecret() };
+    } catch (error) {
+      throw toAppError(error, ErrorService.Locks, 'LocksService.exchangeSessionCode');
+    }
+  }
+
+  /** Restores a Locks session from a persisted bearer secret. */
+  static async restoreSession({ lockServerPubky, secret }: TRestoreLocksSessionParams): Promise<LocksSdkSession> {
+    try {
+      await ensureLocksSdkReady();
+      return this.buildLocksClient(lockServerPubky).restoreSession(secret);
+    } catch (error) {
+      throw toAppError(error, ErrorService.Locks, 'LocksService.restoreSession');
+    }
+  }
+
+  /** Signs the Locks session out on the Lock Server. */
+  static async signout(session: LocksSdkSession): Promise<void> {
+    try {
+      await session.signout();
+    } catch (error) {
+      throw toAppError(error, ErrorService.Locks, 'LocksService.signout');
+    }
+  }
+
+  /**
+   * Registers the creator's default Lock Server by writing the lock-service pointer
+   * (`/pub/locks.app/config.json`). The Lock Server performs the homeserver write; the SDK session
+   * carries the frontend-session bearer, so the FE attaches no `Authorization` header by hand.
+   */
+  static async setLockServiceConfig(session: LocksSdkSession, defaultLockServer: string): Promise<void> {
+    try {
+      await ensureLocksSdkReady();
+      await session.creator.setLockServicePointer(new SetLockServicePointerOptions(defaultLockServer));
+    } catch (error) {
+      throw toAppError(error, ErrorService.Locks, 'LocksService.setLockServiceConfig');
+    }
+  }
 
   /**
    * Uploads one file's raw bytes under `/priv/locks.app/content/<path>` and returns its descriptor
@@ -61,7 +159,7 @@ export class LocksContentService {
       )) as { creator: string; guarded_resource: TGuardedResource };
       return { resource: response.guarded_resource, creator: response.creator };
     } catch (error) {
-      throw toLocksError(error, 'LocksContentService.registerGuardedResource');
+      throw toLocksError(error, 'LocksService.registerGuardedResource');
     }
   }
 
@@ -116,7 +214,7 @@ export class LocksContentService {
         creator: response.content_lock.creator,
       };
     } catch (error) {
-      throw toLocksError(error, 'LocksContentService.createContentLock');
+      throw toLocksError(error, 'LocksService.createContentLock');
     }
   }
 }
